requirements.txt
streamlit
pandas
requests
PyPDF2
python-docx
openai
gspread
oauth2client
